remove automagic command path fixing misfeature
Qmake tried to extract the actual executable part of an
extra compiler's commands and depend_command value and
then "fix" it by replacing the directory separators in
it with their local versions and calling QDir::cleanPath
on it.

This misfeature was implemented incompletely and led to
unexpected results (see the numerous attempts to fix
QTBUG-16372).

The user is responsible for passing a correct command by
calling the shell_quote or shell_escape functions if
necessary.

Change-Id: Ic4bfe9eeb697775cd99c865e7a9d335e63605dea
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>