qdoc: Now lists variable in namespaces
qdoc never did list variables declared in a namespace. It was
probably an oversight, because there are currently no uses of
the \variable command in namespaces. But recently a developer
tried to use \variable to document a namespace variable, and
it didn't work. This update corrects that problem.

Task-number: QTBUG-38734
Change-Id: I47b6234f2bafbb73dfb755a2ad82c9596a4489b1
Reviewed-by: Jørgen Lind <jorgen.lind@digia.com>
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>