QRegularExpression: allow users to skip the UTF-16 check of the subject string
PCRE does not handle invalid UTF-16 sequences. For this reason we always
check a subject string's UTF-16 validity before attempting any match
over it (actually we let PCRE do that).

The only exception so far has been global matching -- once the first
match was done, we skipped re-doing the check over and over again the
same string (PCRE actually checks the /entire/ string, not only the part
it uses for matching).

Still, users had no way to skip this check if they were 100% sure the
string was a valid UTF-16 string. This commit introduces a way for them
to skip the check.

Change-Id: Iea352c06f531aa2153863b3a1681acaab7ac375c
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>