Partially revert "Cocoa: support modal windows"
This reverts parts of commit d9875f7bff6d52a52a1d0bf4002044a5304cf6bf,
in particular the code for "2. Make interrupt() use [NSApp abortModal]"

abortModal is not the right way to end a modal session,
and introduced bad side effects, as reported in
QTBUG-34677.

Restore this part of the event dispatcher to the
Qt 4 state.

Change-Id: Iacc2d4a0757807c87c4320c93ed4db186622945c
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>