Fix multisampled fbo readback
Now that the current fbo is not tracked anymore, toImage() and
blitFramebuffer() will not restore the previously bound (or
whichever Qt thinks was current before) fbo. toImage() needs to
accommodate this since the isBound() check was going wrong now that
the blit always makes the fbo non-current.

The current fbo is now queried via glGet both in toImage() and
blitFramebuffer().

This will fix the recently introduced Lancelot failures with -glbuffer.

Change-Id: I43a780beaeac4697d92cb0ebda7d14ca28a3924f
Reviewed-by: aavit <eirik.aavitsland@digia.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>